--- src/fr/cnam/smb111/cours02/tp2_3/ClientParameters.java
@@ -1,21 +1,0 @@
-package fr.cnam.smb111.cours02.tp2_3;
-
-public class ClientParameters {
-    //Client Commands
-    public static final String DOWN = "DOWN";
-    public static final String UP = "UP";
-    public static final String GET = "GET";
-    public static final String END = "END";
-
-    /* To Clean
-    //Marchaling/Unmarshalling params
-    public static final String MARSHALLING_DELIMETER = "\t";
-
-
-     */
-    /* To Clean
-    //Client communication infos
-    public static final int CLIENT_PORT = 2001;
-    */
-
-}
