--- src/fr/cnam/smb111/cours02/tp2_3/ServerParametersTp2_3.java
@@ -1,6 +1,0 @@
-package fr.cnam.smb111.cours02.tp2_3;
-
-public class ServerParametersTp2_3 {
-    public static final int SERVER_PORT_NUMBER = 2000;
-    //To Clean public static final String SERVER_ADDRESS = null;
-}
