+++ src/fr/cnam/smb111/cours02/tp2_3/client/ClientLauncherMainTest.java
@@ -1,6 +1,6 @@
-package fr.cnam.smb111.cours02.tp2_3;
+package fr.cnam.smb111.cours02.tp2_3.client;
 
-public class ClientLauncherTest {
+public class ClientLauncherMainTest {
     public static void main(String[] args) {
         ClientLauncherTp2_3.start();
     }
