+++ src/fr/cnam/smb111/cours02/tp2_3/client/ClientParameters.java
@@ -1,0 +1,9 @@
+package fr.cnam.smb111.cours02.tp2_3.client;
+
+public class ClientParameters {
+    //Client Commands
+    public static final String DOWN = "DOWN";
+    public static final String UP = "UP";
+    public static final String GET = "GET";
+    public static final String END = "END";
+}
