+++ src/fr/cnam/smb111/cours02/tp2_3/client/CounterClient.java
@@ -1,6 +1,9 @@
-package fr.cnam.smb111.cours02.tp2_3;
+package fr.cnam.smb111.cours02.tp2_3.client;
 
-import fr.cnam.smb111.cours02.tp2_3.textformating.TextColor;
+import fr.cnam.smb111.cours02.tp2_3.DebugTp2_3;
+import fr.cnam.smb111.cours02.tp2_3.client.menu.Menu;
+import fr.cnam.smb111.cours02.tp2_3.client.menu.MenuImpl;
+import fr.cnam.smb111.cours02.tp2_3.client.textformating.TextColor;
 
 import java.io.*;
 import java.net.Socket;
