+++ src/fr/cnam/smb111/cours02/tp2_3/client/menu/Menu.java
@@ -1,4 +1,4 @@
-package fr.cnam.smb111.cours02.tp2_3;
+package fr.cnam.smb111.cours02.tp2_3.client.menu;
 
 public interface Menu {
     void printMenu();
