+++ src/fr/cnam/smb111/cours02/tp2_3/client/textformating/TextColor.java
@@ -1,4 +1,4 @@
-package fr.cnam.smb111.cours02.tp2_3.textformating;
+package fr.cnam.smb111.cours02.tp2_3.client.textformating;
 
 public enum TextColor {
