+++ src/fr/cnam/smb111/cours02/tp2_3/server/ServerParametersTp2_3.java
@@ -1,0 +1,5 @@
+package fr.cnam.smb111.cours02.tp2_3.server;
+
+public class ServerParametersTp2_3 {
+    public static final int SERVER_PORT_NUMBER = 2000;
+}
